move some MIR const pretty-printing into pretty.rs
